Fix upvalue indices
I had misunderstood how "index" was used in upvalues (both compiler and interpreter side).  Turns out that the "index" is either an index into the stack (as a local) or an index into an upvalue list in the parent.

Future misunderstandings can probably be fixed by use of some types for these different indices and some strong comments on the interpreter side that upvalue indices are always parent-relative, rather than absolute.