Fix crash due to fiber being null
stack_trace_fiber is a HeapHandle& into globals.fiber.  However the match:
            match &scope.as_ref(&self.globals).fiber {

Appears to move out of that HeapHandle (not entirely sure why, I would expect the & to prevent that), and cause the stack_trace_fiber HeapHandle& to end up pointing to a null handle.

This shouldn't be possible, and there is likely more work to do to remove this footgun.  But this at least fixes the crash.