Start preparing for collection
vmgc was just taught how to collect on allocate, this is the corresponding changes to safe_wren to make that possible.

It doesn't actually work yet.  For some reason we're failing to allocate in the new space during collection, but will solve that in a later patch.

This also caught two more self recursions (one in safe_wren with Stack::trace and another in vmgc with GCError Display).